rest-one

+++ app/Http/Controllers/Admin/CategoryController.php
@@ -5,7 +5,7 @@ namespace App\Http\Controllers\Admin;
 use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 use App\DataTables\CategoryDataTable;
-use App\Models\Category;
+use App\Models\{Category, PortfolioItem};
 use Illuminate\Support\Str;
 
 class CategoryController extends Controller
@@ -81,6 +81,11 @@ class CategoryController extends Controller
     public function destroy($id)
     {
         $category = Category::findOrFail($id);
-        $category->delete();
+        $hasItem = PortfolioItem::where('category_id', $category->id)->count();
+        if ($hasItem == 0) {
+            $category->delete();
+            return true;
+        }
+        return response(['status'=>'error']);
     }
 }
